requirements.txt
black==25.1.0
coverage==7.9.2
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt-5.5.1
drf-spectacular==0.28.0
flake8==7.3.0
ipython==9.4.0
pillow==11.3.0
psycopg2==2.9.10
python-dotenv==1.1.1
redis==6.2.0
django-cors-headers==4.7.0
